separated out the lists on the sidebar for modularity. made Components/Nav directory.
