+++ src/Components/Nav/Nav.js
@@ -17,13 +17,14 @@ import {
 	ListItemIcon,
 	Typography,
 } from "@mui/material";
-import Color from "./Color.js";
+import Color from "../Color.js";
 import React, { useState } from "react";
 import { Routes, Route, useNavigate, Link } from "react-router-dom";
 import InboxIcon from "@mui/icons-material/MoveToInbox";
 import MailIcon from "@mui/icons-material/Mail";
 import MenuIcon from "@mui/icons-material/Menu";
 import { Logout } from "@mui/icons-material";
+import { sidebarList1, sidebarList2 } from "./SidebarLists.js";
 
 export default function Nav({ themes, activeTheme, updateActiveTheme, token, setToken }) {
 	const [sidebar, updateSidebar] = useState(false);
@@ -50,29 +51,9 @@ export default function Nav({ themes, activeTheme, updateActiveTheme, token, set
 						onKeyDown={toggleDrawer(false)}
 						style={{ backgroundColor: Color(2, activeTheme, themes) }}
 					>
-						<List>
-							{["Inbox", "Starred", "Send email", "Drafts"].map(
-								(text, index) => (
-									<ListItem button key={text}>
-										<ListItemIcon>
-											{index % 2 === 0 ? <InboxIcon /> : <MailIcon />}
-										</ListItemIcon>
-										<ListItemText primary={text} />
-									</ListItem>
-								)
-							)}
-						</List>
+						{sidebarList1()}	
 						<Divider />
-						<List>
-							{["All mail", "Trash", "Spam"].map((text, index) => (
-								<ListItem button key={text}>
-									<ListItemIcon>
-										{index % 2 === 0 ? <InboxIcon /> : <MailIcon />}
-									</ListItemIcon>
-									<ListItemText primary={text} />
-								</ListItem>
-							))}
-						</List>
+						{sidebarList2()}
 						<Select
 							labelId="demo-simple-select-label"
 							id="demo-simple-select"
