+++ src/Components/Nav/SidebarLists.js
@@ -1,0 +1,58 @@
+import {
+	AppBar,
+	Autocomplete,
+	Box,
+	Button,
+	Card,
+	Drawer,
+	IconButton,
+	InputLabel,
+	MenuItem,
+	Select,
+	List,
+	ListItem,
+	Divider,
+	ListItemText,
+	Toolbar,
+	ListItemIcon,
+	Typography,
+} from "@mui/material";
+import Color from "../Color.js";
+import React, { useState } from "react";
+import { Routes, Route, useNavigate, Link } from "react-router-dom";
+import InboxIcon from "@mui/icons-material/MoveToInbox";
+import MailIcon from "@mui/icons-material/Mail";
+import MenuIcon from "@mui/icons-material/Menu";
+import { Logout } from "@mui/icons-material";
+
+export function sidebarList1 () {
+    return (
+        <List>
+            {["Inbox", "Starred", "Send email", "Drafts"].map(
+                (text, index) => (
+                    <ListItem button key={text}>
+                        <ListItemIcon>
+                            {index % 2 === 0 ? <InboxIcon /> : <MailIcon />}
+                        </ListItemIcon>
+                        <ListItemText primary={text} />
+                    </ListItem>
+                )
+            )}
+        </List>
+    )
+}
+
+export function sidebarList2 () {
+    return (
+        <List>
+            {["All mail", "Trash", "Spam"].map((text, index) => (
+                <ListItem button key={text}>
+                    <ListItemIcon>
+                        {index % 2 === 0 ? <InboxIcon /> : <MailIcon />}
+                    </ListItemIcon>
+                    <ListItemText primary={text} />
+                </ListItem>
+            ))}
+        </List>
+    )
+}
